Se ajusta logica en el componente ConsultRawMaterialsOff para no afectar el componente padre ConsultRawMateralsPage al mostrar los registros dados de baja
