+++ src/components/Platform/03Inventories/RawMaterials/01ConsultRawMaterialsOff/styles.module.css
