+++ src/components/Platform/03Inventories/RawMaterials/06ModalRawMaterialOff/styles.module.css
