+++ src/pages/Platform/PanelUser/03Inventories/04InventoryRawMaterals/CreateRawMaterals/CreateRawMateralsPage.tsx
@@ -54,23 +54,9 @@ function CreateRawMateralsPage({ selectedBranchId, onCreateComplete, onRawMateri
         setNameItem(event.target.value);
     };
 
-    //Setea si la materia prima tiene o no empaque principal
-    const [ selectedpackaged, setSelectedpackaged ] = useState('Si');
-    const handlepackagedChange = (value: 'Si' | 'No') => {
-        setSelectedpackaged(value);
-        setValue('packaged', value);
-    };
-
-    //Setea el valor 'Si' o 'No' en la propiedad "returnablePackaging"
-    const [selectedReturnablePackaging, setSelectedReturnablePackaging] = useState('Si');
-    const handleReturnablePackagingChange = (value: 'Si' | 'No') => {
-        setSelectedReturnablePackaging(value);
-        setValue('returnablePackaging', value);
-    };
-
     //Setea si el producto aumentará de forma periódica en el inventario
-    const [ inventoryIncrease, setInventoryIncrease ] = useState('Si');
-    const [ periodicityAutomaticIncrease, setPeriodicityAutomaticInventoryIncrease ] = useState<string | undefined>(undefined);
+    const [inventoryIncrease, setInventoryIncrease] = useState('Si');
+    const [periodicityAutomaticIncrease, setPeriodicityAutomaticInventoryIncrease] = useState<string | undefined>(undefined);
     const handleInventoryIncrease = (value: 'Si' | 'No') => {
         setInventoryIncrease(value);
         setPeriodicityAutomaticInventoryIncrease(undefined)
@@ -82,22 +68,31 @@ function CreateRawMateralsPage({ selectedBranchId, onCreateComplete, onRawMateri
         setPeriodicityAutomaticInventoryIncrease(value);
     };
 
+    //Setea si la materia prima está empacada
+    const [selectedpackaged, setSelectedpackaged] = useState('Si');
+    const handlePackagedChange = (value: 'Si' | 'No') => {
+        setSelectedpackaged(value);
+        setValue('packaged', value);
+    };
+
+    //Setea el valor 'Si' o 'No' en la propiedad "returnablePackaging"
+    const [selectedReturnablePackaging, setSelectedReturnablePackaging] = useState('Si');
+    const handleReturnablePackagingChange = (value: 'Si' | 'No') => {
+        setSelectedReturnablePackaging(value);
+        setValue('returnablePackaging', value);
+    };
+
     //Setea la unidad de medida
     const [showUnitMeasure, setShowUnitMeasure] = useState('');
     const handleUnitMeasureChange = (event: { target: { value: SetStateAction<string> }}) => {
         setShowUnitMeasure(event.target.value);
     };
-    
+
     //Setea el valor 'Si' o 'No' en la propiedad "individualPackaging"
     const [selectedIndividualPackaging, setSelectedIndividualPackaging] = useState('Si');
     const handleIndividualPackagingChange = (value: 'Si' | 'No') => {
         setSelectedIndividualPackaging(value);
         setValue('individualPackaging', value);
-    };
-
-    const [showIndividualPackage, setShowIndividualPackage] = useState('');
-    const handleIndividualPackageChange = (event: { target: { value: SetStateAction<string> }}) => {
-        setShowIndividualPackage(event.target.value);
     };
 
     const onSubmit = (values: IRawMaterial) => {
@@ -172,9 +167,10 @@ function CreateRawMateralsPage({ selectedBranchId, onCreateComplete, onRawMateri
                             {formSubmitted && (
                                 <div className={`${styles.alert__Success} text-center position-absolute alert-success`}>El formulario se ha enviado con éxito</div>
                             )}
-                            {errorRawMaterial?.map((error, i) => (
+                            {Array.isArray(errorRawMaterial) && errorRawMaterial?.map((error, i) => (
                                 <div key={i} className={`${styles.alert__Danger} text-center position-absolute alert-danger`}>{error}</div>
                             ))}
+                        
                             <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
                                 <div>
                                     <p className={`${styles.text} mb-0 p-2`}>Selecciona una Sede</p>
@@ -199,6 +195,20 @@ function CreateRawMateralsPage({ selectedBranchId, onCreateComplete, onRawMateri
 
                             <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
                                 <div>
+                                    <p className={`${styles.text} mb-0 p-2`}>La materia primas que vas a registrar ¿Tiene código de barras?</p>
+                                </div>
+                                <div>
+                                    <input
+                                        type="text"
+                                        {...register('barCode')}
+                                        className={`${styles.input} p-2 border `}
+                                        placeholder='Código de barras de la materia prima que quieres registrar'
+                                    />
+                                </div>
+                            </div>
+
+                            <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
+                                <div>
                                 <p className={`${styles.text} mb-0 p-2`} >¿Cuál es el nombre de la materia prima que vas a registrar?</p>
                                 </div>
                                 <div>
@@ -216,7 +226,211 @@ function CreateRawMateralsPage({ selectedBranchId, onCreateComplete, onRawMateri
                             </div>
 
                             <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
-                                <div>                           
+                                <div>
+                                    <p className={`${styles.text} mb-0 p-2`}>¿Cuál es la marca de la materia prima?</p>
+                                </div>
+                                <div>
+                                    <input
+                                        type="text"
+                                        {...register('brandItem')}
+                                        className={`${styles.input} p-2 border `}
+                                        placeholder='Marca de la materia prima que quieres registrar'
+                                    />
+                                </div>
+                            </div>
+
+                            <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
+                                <div>
+                                    <p className={`${styles.text}`} >¿La materia prima/insumo viene empacada?</p>
+                                </div>
+                                <div className={`${styles.conditionContainer} d-flex align-items-center justify-content-center  border rounded`}>
+                                    <div
+                                        className={`${styles.conditionOption} ${selectedpackaged === 'Si' ? styles.selected : ''} m-1 p-2 text-center`}
+                                        onClick={() => handlePackagedChange('Si')}
+                                    >
+                                        Si
+                                    </div>
+                                    <div
+                                        className={`${styles.conditionOption} ${selectedpackaged === 'No' ? styles.selected : ''} m-1 p-2 text-center`}
+                                        onClick={() => handlePackagedChange('No')}
+                                    >
+                                        No
+                                    </div>
+                                    {errors.packaged && (
+                                        <p className='text-danger'>Este dato es requerido</p>
+                                    )}
+                                </div>
+                            </div>
+
+                            {selectedpackaged === 'Si' && (
+                                <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
+                                    <div className="px-3">
+                                        <p className={`${styles.text} mb-0 p-2`} >Si la materia prima viene empacada ¿Cuál es el tipo de empaque principal?</p>
+                                    </div>
+                                    <div>
+                                        <select
+                                            {...register('primaryPackageType')}
+                                            className={`${styles.input} p-2 border `}
+                                        >
+                                            <option value='Ninguno'>Ninguno</option>
+                                            <option value='Papel'>Papel</option>
+                                            <option value='Papel de archivo'>Papel de archivo</option>
+                                            <option value='Carton'>Cartón</option>
+                                            <option value='Aluminio'>Aluminio</option>
+                                            <option value='Plegadiza'>Plegadiza</option>
+                                            <option value='Vidrio'>Vidrio</option>
+                                            <option value='PET / PETE Polietileno Tereftalato'>PET / PETE Polietileno Tereftalato</option>
+                                            <option value='HDPE Polietileno de alta densidad'>HDPE Polietileno de alta densidad</option>
+                                            <option value='PVC Policloruro de Vinilo'>PVC Policloruro de Vinilo</option>
+                                            <option value='LDPE Polietileno de baja densidad'>LDPE Polietileno de baja densidad</option>
+                                            <option value='PP Polipropileno'>PP Polipropileno</option>
+                                            <option value='PS Poliestireno'>PS Poliestireno</option>
+                                            <option value='Otros plasticos (Policarbonato, estireno, nylon)'>Otros plásticos (Policarbonato, estireno, nylon)</option>
+                                            <option value='Hierro'>Hierro</option>
+                                            <option value='Icopor'>Icopor</option>
+                                            <option value='Biodegradable'>Biodegradable</option>
+                                            <option value='Plastico de burbujas'>Plástico de burbujas</option>
+                                        </select>
+                                        {errors.primaryPackageType && (
+                                            <p className='text-danger'>El tipo de empaque de tu materia prima es requerido</p>
+                                        )}
+                                    </div>
+                                </div>
+                            )}
+
+                            <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
+                                <div>
+                                    <p className={`${styles.text} mb-0 p-2`} >¿La materia prima tiene empaques adicionales?</p>
+                                </div>
+                                <div className={`${styles.conditionContainer} d-flex align-items-center justify-content-center border rounded`}>
+                                    <div
+                                        className={`${styles.conditionOption} ${selectedIndividualPackaging === 'Si' ? styles.selected : ''} m-1 p-2 text-center`}
+                                        onClick={() => handleIndividualPackagingChange('Si')}
+                                    >
+                                        Si
+                                    </div>
+                                    <div
+                                        className={`${styles.conditionOption} ${selectedIndividualPackaging === 'No' ? styles.selected : ''} m-1 p-2 text-center`}
+                                        onClick={() => handleIndividualPackagingChange('No')}
+                                    >
+                                        No
+                                    </div>
+                                    {errors.individualPackaging && (
+                                        <p className='text-danger'>Este dato es requerido</p>
+                                    )}
+                                </div>
+                            </div>
+
+                            {selectedIndividualPackaging === 'Si' && (
+                                <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
+                                    <div>
+                                        <p className={`${styles.text} mb-0 p-2`} >Si la materi prima tiene empaques adicionales ¿Cuál es el tipo de empaque?</p>
+                                    </div>
+                                    <div>
+                                        <select
+                                            {...register('secondaryPackageType', { required: true })}
+                                            className={`${styles.input} p-2 border `}                                    
+                                        >
+                                            <option value='Papel'>Papel</option>
+                                            <option value='Papel de archivo'>Papel de archivo</option>
+                                            <option value='Carton'>Cartón</option>                                                
+                                            <option value='Aluminio'>Aluminio</option>
+                                            <option value='Plegadiza'>Plegadiza</option>
+                                            <option value='Vidrio'>Vidrio</option>
+                                            <option value='PET / PETE Polietileno Tereftalato'>PET / PETE Polietileno Tereftalato</option>                                                
+                                            <option value='HDPE Polietileno de alta densidad'>HDPE Polietileno de alta densidad</option>
+                                            <option value='PVC Policloruro de Vinilo'>PVC Policloruro de Vinilo</option>
+                                            <option value='LDPE Polietileno de baja densidad'>LDPE Polietileno de baja densidad</option>
+                                            <option value='PP Polipropileno'>PP Polipropileno</option>
+                                            <option value='PS Poliestireno'>PS Poliestireno</option>
+                                            <option value='Otros plasticos (Policarbonato, estireno, nylon)'>Otros plásticos (Policarbonato, estireno, nylon)</option>
+                                            <option value='Hierro'>Hierro</option>
+                                            <option value='Icopor'>Icopor</option>
+                                            <option value='Biodegradable'>Biodegradable</option>
+                                            <option value='Plastico de burbujas'>Plástico de burbujas</option>
+                                        </select>
+                                        {errors.secondaryPackageType && (
+                                            <p className='text-danger'>El tipo de empaque de tu materi aprima es requerido</p>
+                                        )}
+                                    </div>
+                                </div>
+                            )}
+
+                            {selectedpackaged === 'Si' && (
+                                <div>
+                                    <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
+                                        <div className="px-3">
+                                            <p className={`${styles.text} mb-0 p-2`} >¿Cuánt{['Unidades', 'Onza', 'Pimpina', 'Libra', 'Arroba', 'Tonelada'].includes(showUnitMeasure) ? 'as' : 'os'} {showUnitMeasure}{['Unidades'].includes(showUnitMeasure) ? '' : 's'} de "{nameItem}" vienen por cada empaque o paquete?</p>
+                                        </div>
+                                        <div>
+                                            <input
+                                                type="number"
+                                                {...register('quantityPerPackage', { required: true, setValueAs: (value) => parseFloat(value) })}
+                                                className={`${styles.input} p-2 border `}
+                                                placeholder='Ej: 10'
+                                                min={0}
+                                                onKeyDown={(e) => {
+                                                    if (e.key === '-' || e.key === 'e' || e.key === '+' || e.key === '.') {
+                                                        e.preventDefault();
+                                                    }
+                                                }}
+                                            />
+                                            {errors.quantityPerPackage && (
+                                                <p className='text-danger'>El valor en {showUnitMeasure} es requerido</p>
+                                            )}
+                                        </div>
+                                    </div>
+
+                                    <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
+                                        <div>
+                                            <p className={`${styles.text} mb-0 p-2`} >¿El empaque, embalaje o envoltura de tu mercancía es retornable?</p>
+                                        </div>
+                                        <div className={`${styles.conditionContainer} d-flex align-items-center justify-content-center  border rounded`}>
+                                            <div
+                                                className={`${styles.conditionOption} ${selectedIndividualPackaging === 'Si' ? styles.selected : ''} m-1 p-2 text-center`}
+                                                onClick={() => handleReturnablePackagingChange('Si')}
+                                            >
+                                                Si
+                                            </div>
+                                            <div
+                                                className={`${styles.conditionOption} ${selectedIndividualPackaging === 'No' ? styles.selected : ''} m-1 p-2 text-center`}
+                                                onClick={() => handleReturnablePackagingChange('No')}
+                                            >
+                                                No
+                                            </div>
+                                            {errors.returnablePackaging && (
+                                                <p className='text-danger'>Este dato es requerido</p>
+                                            )}
+                                        </div>
+                                    </div>
+                                </div>
+                            )}
+
+                            <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
+                                <div>
+                                    <p className={`${styles.text} mb-0 p-2`} >Hoy siendo la primer vez que registras información, ¿Cuánta materia prima tienes en el inventario?</p>
+                                </div>
+                                <div>
+                                    <input
+                                        type="number"
+                                        {...register('inventory', { required: true, setValueAs: (value) => parseFloat(value) })}
+                                        className={`${styles.input} p-2 border `}
+                                        placeholder='Tu inventario acá'
+                                        min={0}
+                                        onKeyDown={(e) => {
+                                            if (e.key === '-' || e.key === 'e' || e.key === '+' || e.key === '.') {
+                                                e.preventDefault();
+                                            }
+                                        }}
+                                    />
+                                    {errors.inventory && (
+                                        <p className='text-danger'>El inventario de la materia prima es requerido</p>
+                                    )}
+                                </div>
+                            </div>
+
+                            <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
+                                <div>
                                     <p className={`${styles.text} mb-0 p-2`} >¿En qué unidad de medida viene la materia prima?</p>
                                 </div>
                                 <div>
@@ -263,29 +477,6 @@ function CreateRawMateralsPage({ selectedBranchId, onCreateComplete, onRawMateri
                                     </select>
                                     {errors.unitMeasure && (
                                         <p className='text-danger'>El tipo de empaque de tu materia prima es requerido</p>
-                                    )}
-                                </div>
-                            </div>
-
-                            <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
-                                <div className="px-3">
-                                    <p className={`${styles.text} mb-0 p-2`} >Inventario: ¿Cuánta materia prima tienes en inventario?</p>
-                                </div>
-                                <div>
-                                    <input
-                                        type="number"
-                                        {...register('inventory', { required: true, setValueAs: (value) => parseFloat(value) })}
-                                        className={`${styles.input} p-2 border `}
-                                        placeholder='Tu inventario acá'
-                                        min={0}
-                                        onKeyDown={(e) => {
-                                            if (e.key === '-' || e.key === 'e' || e.key === '+' || e.key === '.') {
-                                                e.preventDefault();
-                                            }
-                                        }}
-                                    />
-                                    {errors.inventory && (
-                                        <p className='text-danger'>El inventario de la materia prima es requerido</p>
                                     )}
                                 </div>
                             </div>
@@ -395,224 +586,43 @@ function CreateRawMateralsPage({ selectedBranchId, onCreateComplete, onRawMateri
 
                             <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
                                 <div>
-                                    <p className={`${styles.text}`} >¿La materia prima/insumo viene empacada?</p>
+                                    <p className={`${styles.text} mb-0 p-2`} >¿Cuál es el precio de compra antes de impuestos?</p>
                                 </div>
-                                <div className={`${styles.conditionContainer} d-flex align-items-center justify-content-center  border rounded`}>
-                                    <div
-                                        className={`${styles.conditionOption} ${selectedpackaged === 'Si' ? styles.selected : ''} m-1 p-2 text-center`}
-                                        onClick={() => handlepackagedChange('Si')}
-                                    >
-                                        Si
-                                    </div>
-                                    <div
-                                        className={`${styles.conditionOption} ${selectedpackaged === 'No' ? styles.selected : ''} m-1 p-2 text-center`}
-                                        onClick={() => handlepackagedChange('No')}
-                                    >
-                                        No
-                                    </div>
-                                    {errors.packaged && (
-                                        <p className='text-danger'>Este dato es requerido</p>
+                                <div>
+                                    <input
+                                        type="number"
+                                        {...register('purchasePriceBeforeTax', { required: true })}
+                                        className={`${styles.input} p-2 border `}
+                                        placeholder='Precio de la materia prima'
+                                        min={0}
+                                        onKeyDown={(e) => {
+                                            if (e.key === '-' || e.key === 'e' || e.key === '+' || e.key === '.') {
+                                                e.preventDefault();
+                                            }
+                                        }}
+                                    />
+                                    {errors.purchasePriceBeforeTax && (
+                                        <p className='text-danger'>El el precio de compra antes de impuestos es requerido</p>
                                     )}
                                 </div>
                             </div>
-
-                            {selectedpackaged === 'Si' && (
-                                <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
-                                    <div className="px-3">
-                                        <p className={`${styles.text} mb-0 p-2`} >Si la materia prima viene empacada ¿Cuál es el tipo de empaque principal?</p>
-                                    </div>
-                                    <div>
-                                        <select
-                                            {...register('primaryPackageType')}
-                                            className={`${styles.input} p-2 border `}
-                                        >
-                                            <option value='Ninguno'>Ninguno</option>
-                                            <option value='Papel'>Papel</option>
-                                            <option value='Papel de archivo'>Papel de archivo</option>
-                                            <option value='Carton'>Cartón</option>
-                                            <option value='Aluminio'>Aluminio</option>
-                                            <option value='Plegadiza'>Plegadiza</option>
-                                            <option value='Vidrio'>Vidrio</option>
-                                            <option value='PET / PETE Polietileno Tereftalato'>PET / PETE Polietileno Tereftalato</option>
-                                            <option value='HDPE Polietileno de alta densidad'>HDPE Polietileno de alta densidad</option>
-                                            <option value='PVC Policloruro de Vinilo'>PVC Policloruro de Vinilo</option>
-                                            <option value='LDPE Polietileno de baja densidad'>LDPE Polietileno de baja densidad</option>
-                                            <option value='PP Polipropileno'>PP Polipropileno</option>
-                                            <option value='PS Poliestireno'>PS Poliestireno</option>
-                                            <option value='Otros plasticos (Policarbonato, estireno, nylon)'>Otros plásticos (Policarbonato, estireno, nylon)</option>
-                                            <option value='Hierro'>Hierro</option>
-                                            <option value='Icopor'>Icopor</option>
-                                            <option value='Biodegradable'>Biodegradable</option>
-                                            <option value='Plastico de burbujas'>Plástico de burbujas</option>
-                                        </select>
-                                        {errors.primaryPackageType && (
-                                            <p className='text-danger'>El tipo de empaque de tu materia prima es requerido</p>
-                                        )}
-                                    </div>
-                                </div>
-                            )}
-
-                            {selectedpackaged === 'Si' && (
-                                <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
-                                    <div>
-                                        <p className={`${styles.text} mb-0 p-2`} >¿El empaque, embalaje o envoltura de tu materia prima es retornable?</p>
-                                    </div>
-                                    <div className={`${styles.conditionContainer} d-flex align-items-center justify-content-center  border rounded`}>
-                                        <div
-                                            className={`${styles.conditionOption} ${selectedReturnablePackaging === 'Si' ? styles.selected : ''} m-1 p-2 text-center`}
-                                            onClick={() => handleReturnablePackagingChange('Si')}
-                                            >
-                                            Si
-                                        </div>
-                                        <div
-                                            className={`${styles.conditionOption} ${selectedReturnablePackaging === 'No' ? styles.selected : ''} m-1 p-2 text-center`}
-                                            onClick={() => handleReturnablePackagingChange('No')}
-                                        >
-                                            No
-                                        </div>
-                                        {errors.returnablePackaging && (
-                                            <p className='text-danger'>Este dato es requerido</p>
-                                        )}
-                                    </div>
-                                </div>
-                            )}
-
-                            {showUnitMeasure && (
-                                <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
-                                    <div className="px-3">                                    
-                                        <p className={`${styles.text} mb-0 p-2`} >¿Cuánt{['Unidades', 'Onza', 'Pimpina', 'Libra', 'Arroba', 'Tonelada'].includes(showUnitMeasure) ? 'as' : 'os'} {showUnitMeasure}{['Unidades'].includes(showUnitMeasure) ? '' : 's'} de "{nameItem}" vienen por cada empaque o paquete?</p>
-                                    </div>
-                                    <div>
-                                        <input
-                                            type="number"
-                                            {...register('quantityPerPackage', { required: true, setValueAs: (value) => parseFloat(value) })}
-                                            className={`${styles.input} p-2 border `}
-                                            placeholder='Ej: 10'
-                                            min={0}
-                                            onKeyDown={(e) => {
-                                                if (e.key === '-' || e.key === 'e' || e.key === '+' || e.key === '.') {
-                                                    e.preventDefault();
-                                                }
-                                            }}
-                                        />
-                                        {errors.quantityPerPackage && (
-                                            <p className='text-danger'>El valor en {showUnitMeasure} es requerido</p>
-                                        )}
-                                    </div>
-                                </div>
-                            )}
 
                             <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
-                                <div>
-                                    <p className={`${styles.text} mb-0 p-2`} >¿La materia prima tiene empaques adicionales?</p>
+                                <div className="px-3">
+                                    <p className={`${styles.text} mb-0 p-2`} >¿Cuál es el IVA de la materia prima?</p>
                                 </div>
-                                <div className={`${styles.conditionContainer} d-flex align-items-center justify-content-center  border rounded`}>
-                                    <div
-                                        className={`${styles.conditionOption} ${selectedIndividualPackaging === 'Si' ? styles.selected : ''} m-1 p-2 text-center`}
-                                        onClick={() => handleIndividualPackagingChange('Si')}
+                                <div className={styles.containerInput}>
+                                    <select
+                                        defaultValue={'0'}
+                                        className={`${styles.input} p-2 border `}
+                                        {...register('IVA', { required: true })}
                                     >
-                                        Si
-                                    </div>
-                                    <div
-                                        className={`${styles.conditionOption} ${selectedIndividualPackaging === 'No' ? styles.selected : ''} m-1 p-2 text-center`}
-                                        onClick={() => handleIndividualPackagingChange('No')}
-                                    >
-                                        No
-                                    </div>
-                                    {errors.returnablePackaging && (
-                                        <p className='text-danger'>Este dato es requerido</p>
-                                    )}
+                                        <option value='0'>0 %</option>
+                                        <option value='5'>5 %</option>
+                                        <option value='19'>19 %</option>
+                                    </select>
                                 </div>
                             </div>
-
-                            {selectedIndividualPackaging === 'Si' && (
-                                <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
-                                    <div>
-                                        <p className={`${styles.text} mb-0 p-2`} >Si la materia prima tiene empaques adicionales ¿Cuál es el tipo de empaque?</p>
-                                    </div>
-                                    <div>
-                                        <select
-                                            {...register('secondaryPackageType', { required: true })}
-                                            className={`${styles.input} p-2 border `}                                  
-                                        >
-                                            <option value='Papel'>Papel</option>
-                                            <option value='Papel de archivo'>Papel de archivo</option>
-                                            <option value='Carton'>Cartón</option>                                                
-                                            <option value='Aluminio'>Aluminio</option>
-                                            <option value='Plegadiza'>Plegadiza</option>
-                                            <option value='Vidrio'>Vidrio</option>
-                                            <option value='PET / PETE Polietileno Tereftalato'>PET / PETE Polietileno Tereftalato</option>                                                
-                                            <option value='HDPE Polietileno de alta densidad'>HDPE Polietileno de alta densidad</option>
-                                            <option value='PVC Policloruro de Vinilo'>PVC Policloruro de Vinilo</option>
-                                            <option value='LDPE Polietileno de baja densidad'>LDPE Polietileno de baja densidad</option>
-                                            <option value='PP Polipropileno'>PP Polipropileno</option>
-                                            <option value='PS Poliestireno'>PS Poliestireno</option>
-                                            <option value='Otros plasticos (Policarbonato, estireno, nylon)'>Otros plásticos (Policarbonato, estireno, nylon)</option>
-                                            <option value='Hierro'>Hierro</option>
-                                            <option value='Icopor'>Icopor</option>
-                                            <option value='Biodegradable'>Biodegradable</option>
-                                            <option value='Plastico de burbujas'>Plástico de burbujas</option>
-                                        </select>
-                                        {errors.secondaryPackageType && (
-                                            <p className='text-danger'>El tipo de empaque de tu mercancía es requerido</p>
-                                        )}
-                                    </div>
-                                </div>
-                            )}
-
-                            {showUnitMeasure === 'Unidades' && (
-                                <div>
-                                    <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
-                                        <div className="px-3">
-                                            <p className={`${styles.text} mb-0 p-2`} >¿Las unidades vienen en empaque individual? Si la respuesta es "Si", selecciona la casilla</p>
-                                        </div>
-                                        <div>
-                                            <select
-                                                {...register('individualPackaging', { required: true })}
-                                                className={`${styles.info} p-2 border rounded border-secundary`}                                            
-                                                onChange={handleIndividualPackageChange}
-                                            >
-                                                <option value='Si'>Si</option>
-                                                <option value='No'>No</option>
-                                            </select>
-                                        </div>
-                                    </div>
-                                    {showIndividualPackage && (
-                                        <div className="mb-3 p-2 d-flex align-items-center justify-content-center border rounded">
-                                            <div className="px-3">
-                                                <p className={`${styles.text} mb-0 p-2`} >¿Cuál es el tipo de empaque de cada unidad?</p>
-                                            </div>
-                                            <div>
-                                                <select
-                                                    {...register('secondaryPackageType', { required: true })}
-                                                    className={`${styles.info} p-2 border rounded border-secundary`}                                            
-                                                >
-                                                    <option value='Papel'>Papel</option>
-                                                    <option value='Papel de archivo'>Papel de archivo</option>
-                                                    <option value='Carton'>Cartón</option>                                                
-                                                    <option value='Aluminio'>Aluminio</option>
-                                                    <option value='Plegadiza'>Plegadiza</option>
-                                                    <option value='Vidrio'>Vidrio</option>
-                                                    <option value='PET / PETE Polietileno Tereftalato'>PET / PETE Polietileno Tereftalato</option>
-                                                    <option value='HDPE Polietileno de alta densidad'>HDPE Polietileno de alta densidad</option>
-                                                    <option value='PVC Policloruro de Vinilo'>PVC Policloruro de Vinilo</option>
-                                                    <option value='LDPE Polietileno de baja densidad'>LDPE Polietileno de baja densidad</option>
-                                                    <option value='PP Polipropileno'>PP Polipropileno</option>
-                                                    <option value='PS Poliestireno'>PS Poliestireno</option>
-                                                    <option value='Otros plasticos (Policarbonato, estireno, nylon)'>Otros plásticos (Policarbonato, estireno, nylon)</option>
-                                                    <option value='Hierro'>Hierro</option>
-                                                    <option value='Icopor'>Icopor</option>
-                                                    <option value='Biodegradable'>Biodegradable</option>
-                                                    <option value='Plastico de burbujas'>Plástico de burbujas</option>
-                                                </select>
-                                                {errors.secondaryPackageType && (
-                                                    <p className='text-danger'>El tipo de empaque de tu materia prima es requerido</p>
-                                                )}
-                                            </div>
-                                        </div>
-                                    )}
-                                </div>
-                            )}
 
                             <div className="mb-4 d-flex align-items-center justify-content-center">
                                 <button type='submit' className={`${styles.button__Submit} border-0 rounded text-decoration-none`} >Enviar</button>
