+++ src/types/User/rawMaterial.types.ts
@@ -7,28 +7,31 @@ export interface InventoryOffItem {
 
 export interface IRawMaterial {
     id: string;
-    nameItem: string;
     barCode?: string;
+    nameItem: string;
+    brandItem?: string;
+    packaged: 'Si' | 'No';
+    primaryPackageType?: 'Ninguno' | 'Papel' | 'Papel de archivo' | 'Carton' | 'Aluminio' | 'Plegadiza' | 'Vidrio' | 'PET / PETE Polietileno Tereftalato' | 'HDPE Polietileno de alta densidad' | 'PVC Policloruro de Vinilo' | 'LDPE Polietileno de baja densidad' | 'PP Polipropileno' | 'PS Poliestireno' | 'Otros plasticos (Policarbonato, estireno, nylon)' | 'Hierro' | 'Icopor' | 'Biodegradable' | 'Plastico de burbujas';
+    individualPackaging?: 'Si' | 'No';
+    secondaryPackageType?: 'Ninguno' | 'Papel' | 'Papel de archivo' | 'Carton' | 'Aluminio' | 'Plegadiza' | 'Vidrio' | 'PET / PETE Polietileno Tereftalato' | 'HDPE Polietileno de alta densidad' | 'PVC Policloruro de Vinilo' | 'LDPE Polietileno de baja densidad' | 'PP Polipropileno' | 'PS Poliestireno' | 'Otros plasticos (Policarbonato, estireno, nylon)' | 'Hierro' | 'Icopor' | 'Biodegradable' | 'Plastico de burbujas';
+    quantityPerPackage?: number;
+    returnablePackaging?: 'Si' | 'No';
     inventory: number;
-    inventoryOff?: InventoryOffItem[];
     unitMeasure: 'Unidades' | 'Ristra' | 'Decena' | 'Docena' | 'Miligramo' | 'Gramo' | 'Media libra' | 'Libra' | 'Kilogramo' | 'Caja' | 'Paca' | 'Arroba' | 'Bulto' | 'Saco' | 'Tonelada' | 'Mililitro' | 'Onza' | 'Litro' | 'Galon' | 'Pimpina' | 'Metro cubico' | 'Milimetro' | 'Centrimetro' | 'Pulgada' | 'Metro' | 'Centimetro cuadrado' | 'Metro cuadrado';
     inventoryIncrease?: 'Si' | 'No';
     periodicityAutomaticIncrease?: 'Diario' | 'Semanal' | 'Quincenal' | 'Mensual' | 'Bimestral' | 'Trimestral' | 'Semestral';
     automaticInventoryIncrease?: number;
-    purchasePriceBeforeTax?: number;
-    IVA?: number;
-    packaged: 'Si' | 'No';
-    primaryPackageType?: 'Ninguno' | 'Papel' | 'Papel de archivo' | 'Carton' | 'Aluminio' | 'Plegadiza' | 'Vidrio' | 'PET / PETE Polietileno Tereftalato' | 'HDPE Polietileno de alta densidad' | 'PVC Policloruro de Vinilo' | 'LDPE Polietileno de baja densidad' | 'PP Polipropileno' | 'PS Poliestireno' | 'Otros plasticos (Policarbonato, estireno, nylon)' | 'Hierro' | 'Icopor' | 'Biodegradable' | 'Plastico de burbujas';
-    expirationDate?: Date;
-    returnablePackaging?: 'Si' | 'No';
-    quantityPerPackage?: number;
-    individualPackaging?: 'Si' | 'No';
-    secondaryPackageType?: 'Ninguno' | 'Papel' | 'Papel de archivo' | 'Carton' | 'Aluminio' | 'Plegadiza' | 'Vidrio' | 'PET / PETE Polietileno Tereftalato' | 'HDPE Polietileno de alta densidad' | 'PVC Policloruro de Vinilo' | 'LDPE Polietileno de baja densidad' | 'PP Polipropileno' | 'PS Poliestireno' | 'Otros plasticos (Policarbonato, estireno, nylon)' | 'Hierro' | 'Icopor' | 'Biodegradable' | 'Plastico de burbujas';
-    inventoryChanges: { date: string; quantity: number, type: 'Ingreso' | 'Salida' }[];
+    purchasePriceBeforeTax: number;
+    IVA: number;
     sellingPrice?: number;
+    isDiscounted?: 'Si' | 'No';
+    discountPercentage?: number;
+    salesCount?: number;
+    expirationDate?: Date;
+    inventoryChanges?: { date: string; quantity: number, type: 'Ingreso' | 'Salida' }[];
+    inventoryOff?: InventoryOffItem[];
     reasonManualDiscountingInventory?: 'Donado' | 'Desechado' | 'Caducado' | 'Perdido' | 'Hurtado';    
     quantityManualDiscountingInventory?: number;    
-    salesCount?: number;
     
     //RELACION CON OTRAS TABLAS
     branchId: string;
